pyproject.toml
[project]
name = "skill-builder"
version = "0.1.0"
description = "ADK agent that uses skills to build skills. Generates agentskills.io-compatible SKILL.md files and exposes itself via A2A protocol for Kagenti deployment."
authors = [{ name = "Raghuram Banda" }]
license = { text = "Apache License 2.0" }
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "google-adk[a2a]>=1.0.0,<2.0.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[dependency-groups]
dev = [
    "google-adk[eval]>=1.0.0",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
